compnents #13 境界越境時に発火するEffectをアタリ付け
