+++ frontend/src/Utilities/GeneEffects/GeneEffectMove.ts
@@ -18,6 +18,12 @@ class GeneEffectMove implements GeneEffectInterface{
     parameter.vector.x += this.parameter.vector.x;
     parameter.vector.y += this.parameter.vector.y;
     parameter.vector.z += this.parameter.vector.z;
+
+    // !!! 仮置き 移動後のポジションの座標を算出する !!!
+    parameter.position.x += this.parameter.vector.x;
+    parameter.position.y += this.parameter.vector.y;
+    parameter.position.z += this.parameter.vector.z;
+    
     return parameter
   }
