+++ frontend/src/Utilities/GlobalVarriables/EffectCatalog.ts
@@ -1,8 +1,7 @@
-import GeneEffectRoll from "../GeneEffects/GeneEffectRoll";
-import GeneEffectMove from "../GeneEffects/GeneEffectMove";
+import GeneEffectRoll from '../GeneEffects/GeneEffectRoll';
+import GeneEffectMove from '../GeneEffects/GeneEffectMove';
 
 export type EffectID = 'ROLL' | 'MOVE';
 export const EFFECT_CATALOG = [
   { id: 'ROLL', effect: GeneEffectRoll },
-  { id: 'MOVE', effect: GeneEffectMove },
-];
+  { id: 'MOVE', effect: GeneEffectMove }];
